pyproject.toml
[project]
name = "log-trace-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-connect>=0.8.17",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
]
